backend/pyproject.toml
[project]
name = "realtime-translate-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "websockets",
    "orjson",
    "httpx",
    "azure-cognitiveservices-speech",
    "python-dotenv",
]

[tool.uv]
dev-dependencies = []
